Removed SCons and Automake build files. Switched license to Apache 2
After testing building with Automake, SCons, Make, and CMake, I have decided to
use CMake.  I have also decided to switch the software license from LGPL3.0 to
Apache 2.0.  Apache 2.0 and LGPL3.0 are very similar, but the LGPL does not play
as well with other licenses; Apache software can be used with LGPL but LGPL
cannot be used in Apache software.